Move Artist definition to base module (gallery) domain package
