+++ app/src/main/java/com/ataulm/artcollector/domain/Artist.kt
@@ -1,0 +1,3 @@
+package com.ataulm.artcollector.domain
+
+data class Artist(val id: String, val name: String)
