--- app/src/main/java/com/ataulm/artcollector/gallery/domain/Artist.kt
@@ -1,3 +1,0 @@
-package com.ataulm.artcollector.gallery.domain
-
-internal data class Artist(val id: String, val name: String)
