+++ app/src/main/java/com/ataulm/artcollector/gallery/domain/Painting.kt
@@ -1,5 +1,7 @@
 package com.ataulm.artcollector.gallery.domain
 
+import com.ataulm.artcollector.domain.Artist
+
 internal data class Painting(
         val id: String,
         val title: String,
