+++ artist/src/main/java/com/ataulm/artcollector/artist/data/AndroidArtistRepository.kt
@@ -4,11 +4,11 @@ import com.ataulm.artcollector.ApiObjectRecord
 import com.ataulm.artcollector.ApiPerson
 import com.ataulm.artcollector.ApiPersonRecord
 import com.ataulm.artcollector.HarvardArtMuseumApi
-import com.ataulm.artcollector.artist.domain.Artist
 import com.ataulm.artcollector.artist.domain.ArtistId
 import com.ataulm.artcollector.artist.domain.ArtistRepository
 import com.ataulm.artcollector.artist.domain.Gallery
 import com.ataulm.artcollector.artist.domain.Painting
+import com.ataulm.artcollector.domain.Artist
 import javax.inject.Inject
 
 internal class AndroidArtistRepository @Inject constructor(
