+++ artist/src/main/java/com/ataulm/artcollector/artist/domain/ArtistRepository.kt
@@ -1,5 +1,7 @@
 package com.ataulm.artcollector.artist.domain
 
+import com.ataulm.artcollector.domain.Artist
+
 internal interface ArtistRepository {
 
     suspend fun artist(): Artist
