+++ artist/src/main/java/com/ataulm/artcollector/artist/ui/ArtistViewModel.kt
@@ -5,16 +5,12 @@ import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.ataulm.artcollector.Event
-import com.ataulm.artcollector.artist.domain.Artist
 import com.ataulm.artcollector.artist.domain.Gallery
 import com.ataulm.artcollector.artist.domain.GetArtistGalleryUseCase
 import com.ataulm.artcollector.artist.domain.GetArtistUseCase
 import com.ataulm.artcollector.artist.domain.Painting
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
+import com.ataulm.artcollector.domain.Artist
+import kotlinx.coroutines.*
 import javax.inject.Inject
 
 internal class ArtistViewModel @Inject constructor(
